create instructorCommentsCommentRow
